Move to top components folder
